Fixed all the merge errors and Tested all the code! MERGE WITH MASTER
