+++ GroupProject01/GroupProject01.sln
@@ -4,8 +4,7 @@ VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GroupProject01", "GroupProject01.csproj", "{80D9EB1C-300C-40E7-9288-685E7B557CB5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NorthwindNUnitOrdersTest", "..\NorthwindNUnitOrdersTest\NorthwindNUnitOrdersTest.csproj", "{2A958CE5-1F00-4530-BCF9-008022C42C44}"
-
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GroupProjectTests", "..\GroupProjectTest\GroupProjectTests.csproj", "{8A5F9961-607C-4044-9793-DEFA917BB3CA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +16,10 @@ Global
 		{80D9EB1C-300C-40E7-9288-685E7B557CB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80D9EB1C-300C-40E7-9288-685E7B557CB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80D9EB1C-300C-40E7-9288-685E7B557CB5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2A958CE5-1F00-4530-BCF9-008022C42C44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2A958CE5-1F00-4530-BCF9-008022C42C44}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2A958CE5-1F00-4530-BCF9-008022C42C44}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2A958CE5-1F00-4530-BCF9-008022C42C44}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A5F9961-607C-4044-9793-DEFA917BB3CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A5F9961-607C-4044-9793-DEFA917BB3CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A5F9961-607C-4044-9793-DEFA917BB3CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A5F9961-607C-4044-9793-DEFA917BB3CA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
